Delete single record

+++ app/Http/Controllers/DataTable/DataTableController.php
@@ -9,7 +9,17 @@ use Illuminate\Database\Eloquent\Builder;
 
 abstract class DataTableController extends Controller
 {
+   /**
+    * If an entity is allowed to be created.
+    * @var boolean
+    */
     protected $allowCreation = true;
+
+    /**
+     * If entity is allowed to be deleted.
+     * @var boolean
+     */
+    protected $allowDeletion = true;
 
 	protected $builder;
 
@@ -37,6 +47,7 @@ abstract class DataTableController extends Controller
     			'records' => $this->getRecords($request),
                 'allow' => [
                     'creation' => $this->allowCreation,
+                    'deletion' => $this->allowDeletion,
                 ],
     		]
     	]);
@@ -58,6 +69,10 @@ abstract class DataTableController extends Controller
     }
 
     public function destroy($id, Request $request) {
+        if (!$this->allowDeletion) {
+            return;
+        }
+
         $this->builder->find($id)->delete();
     }
